added mockup folder for the mock data files and reorganised the structure of css files
