+++ online-shop/src/app/components/navigation/navigation.component.css
@@ -3,7 +3,7 @@
   border-width: thick;
   background-color: #718870;
   color: white;
-  height: 3cm;
+  height: 120px;
   font-family: Verdana, Geneva, Tahoma, sans-serif;
 }
