+++ online-shop/src/app/components/product-details/product-details.component.css
@@ -26,10 +26,6 @@ table {
   padding: auto;
 }
 
-.clearfix {
-  clear: both;
-}
-
 th {
   font-weight: bold;
   text-align: left;
@@ -40,4 +36,8 @@ td {
   text-align: left;
   padding-left: 5%;
   width: 100%;
+}
+
+.clearfix {
+  clear: both;
 }
