+++ online-shop/src/app/mockup/products.ts
